+++ CHANGELOG.md
@@ -16,6 +16,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - N/A
 
+## [3.0.1] 2026-08-29
+Version 3.x added the new template for using photos in orders, instead
+raw text.
+
+### Added
+- product.photo field to the model `src/api/v1/apps/orders/models.py`
+
+### Changed
+- N/A
+
+### Fixed
+- N/A
+
 ## [2.3.4] 2026-08-20
 
 ### Added